+++ src/scenes/Scholarship/ScholarshipShareSaveButtons.js
@@ -9,6 +9,10 @@ import { faFacebookMessenger, faWhatsapp, faFacebook } from '@fortawesome/free-b
 import {toastNotify} from "../../models/Utils";
 import NotificationsService from "../../services/NotificationsService";
 import {Link} from "react-router-dom";
+import {Tooltip} from "antd";
+import {addToMyScholarshipHelper} from "../../models/UserProfile";
+import UserProfileAPI from "../../services/UserProfileAPI";
+import {handleError} from "../../services/utils";
 
 class ScholarshipShareSaveButtons extends React.Component {
 
@@ -32,20 +36,33 @@ class ScholarshipShareSaveButtons extends React.Component {
         const { isSavedScholarship } = this.state;
         const { userProfile, scholarship } = this.props;
 
+        if (!userProfile) {
+            toastNotify((<p>You must <Link to="/register">Register</Link> to save a scholarship.</p>));
+            return;
+        }
+
         if (isSavedScholarship) {
             toastNotify("You've already saved this scholarship 👌🏿");
             return;
         }
 
-        if (!userProfile) {
-            toastNotify((<p>You must <Link to="/register">Register</Link> to save a scholarship.</p>));
-            return;
-        }
+        console.log({userProfile});
+
+        const updatedUserProfile = addToMyScholarshipHelper(userProfile, scholarship);
+        console.log({updatedUserProfile});
         NotificationsService.createScholarshipNotifications(userProfile, scholarship)
             .then(res=> {
                 console.log({res});
-            });
-        this.setState({ isSavedScholarship: !isSavedScholarship });
+                this.setState({isSavedScholarship: !isSavedScholarship});
+                UserProfileAPI
+                    .update({userProfile: updatedUserProfile},
+                        userProfile.user)
+                    .then(res=>{
+                        toastNotify('😃 User Profile successfully saved!');
+                    })
+                    .catch(handleError);
+            })
+            .catch(handleError);
 
     };
 
@@ -105,13 +122,16 @@ class ScholarshipShareSaveButtons extends React.Component {
                         ))}
                     </Dropdown.Menu>
                 </Dropdown>
-                <button className={`btn ${isSavedScholarship ? 'btn-primary' : 'btn-outline-primary'}`}
-                        onClick={this.saveScholarship}
-                        disabled={isSavedScholarship}
-                        title={isSavedScholarship?
-                            "You've already saved this scholarship 👌🏿": 'Save Scholarship'} >
-                    <FontAwesomeIcon className="ml-1" icon={faBookmark}/>
-                </button>
+                <Tooltip placement="right"
+                         title={isSavedScholarship?
+                             "You've already saved this scholarship 👌🏿": 'Save Scholarship'}>
+                    <button className={`btn ${isSavedScholarship ? 'btn-primary' : 'btn-outline-primary'}`}
+                            onClick={this.saveScholarship}
+                            title={isSavedScholarship?
+                                "You've already saved this scholarship 👌🏿": 'Save Scholarship'} >
+                        <FontAwesomeIcon className="ml-1" icon={faBookmark}/>
+                    </button>
+                </Tooltip>
             </div>
         );
     }