+++ src/scenes/Search/Search.js
@@ -3,8 +3,9 @@ import React from 'react';
 import SearchApi from "../../services/SearchAPI";
 import ResponseDisplay from "../../components/ResponseDisplay";
 import {SearchResultsDisplay} from "./SearchResultsDisplay";
-import Form from "react-bootstrap/Form";
 import {Helmet} from "react-helmet";
+import AutoComplete from "../../components/AutoComplete";
+import {MASTER_LIST_EVERYTHING_UNDERSCORE} from "../../models/ConstantsForm";
 
 class Search extends React.Component {
 
@@ -83,11 +84,15 @@ class Search extends React.Component {
 
     updateSearch = event => {
         event.preventDefault();
-        this.setState({searchQuery: event.target.value});
+        this.setState({searchQuery: event.target.value}, () => {
+            this.submitSearch();
+        });
     };
 
     submitSearch = event => {
-        event.preventDefault();
+        if (event && event.preventDefault) {
+            event.preventDefault();
+        }
         const { searchQuery } = this.state;
         this.props.history.push({
             pathname: '/search',
@@ -102,6 +107,9 @@ class Search extends React.Component {
 
         const {isLoadingResponse, responseError, responseOkMessage, searchResults, searchQuery} = this.state;
 
+        const customTheme = {
+            container: 'react-autosuggest__container col-sm-12 col-md-7 p-0 my-3 mx-1',
+        };
         return (
             <div className="container mt-5">
                 <Helmet>
@@ -110,23 +118,22 @@ class Search extends React.Component {
                 </Helmet>
                 <h1>Search {searchQuery && `for ${searchQuery}`}</h1>
 
-                <Form inline  onSubmit={this.submitSearch} className="my-3">
-                    <input value={searchQuery}
-                           className="form-control
-                           search-input
-                           col-sm-12 col-md-7 center-block
-                           mb-3"
-                           type="text"
-                           name="search"
-                           placeholder="Enter a search term" onChange={this.updateSearch}/>
-                    <button className="btn btn-primary col-md-4 col-sm-12 mb-3"
+                <form  onSubmit={this.submitSearch} className="row">
+
+                    <AutoComplete suggestions={MASTER_LIST_EVERYTHING_UNDERSCORE}
+                                  placeholder={"Enter search here"}
+                                  onSelected={this.updateSearch}
+                                  value={searchQuery}
+                                  customTheme={customTheme}
+                                  keyName='search'/>
+                    <button className="btn btn-primary col-md-4 col-sm-12 my-3 mx-1"
                             type="submit">
                         Search
                     </button>
-                </Form>
+                </form>
 
-                { searchResults &&
-                <SearchResultsDisplay searchResults={searchResults} />
+                { searchResults && (searchResults.scholarships || searchResults.essays || searchResults.blogs) &&
+                    <SearchResultsDisplay searchResults={searchResults} />
                 }
                 <ResponseDisplay isLoadingResponse={isLoadingResponse}
                                  responseError={responseError}